chore: renaming SST resources
