+++ infra/api-gateway.ts
@@ -18,13 +18,10 @@ function getDynamicPages(compilation) {
   });
 }
 
+// TODO need to handle basePath here?  (and / or all adapters?)
 const graph = // @ts-expect-error see https://github.com/microsoft/TypeScript/issues/42866
   (await import(new URL("../../public/graph.json", import.meta.url), { with: { type: "json" } }))
     .default;
-// TODO need to handle basePath here?  (and / or all adapters?)
-const apiRoutes = // @ts-expect-error see https://github.com/microsoft/TypeScript/issues/42866
-  (await import(new URL("../../public/manifest.json", import.meta.url), { with: { type: "json" } }))
-    .default.apis.value;
 const ssrPages = getDynamicPages({ config: { prerender: true }, graph });
 
 // https://sst.dev/docs/component/aws/apigatewayv2
@@ -41,17 +38,6 @@ ssrPages.forEach((page) => {
 
   api.route(`GET /routes${routePattern}`, {
     bundle: `.aws-output/routes/${id}`,
-    handler: "index.handler",
-    runtime: RUNTIME,
-  });
-});
-
-apiRoutes.forEach((apiRoute) => {
-  const [route, { id }] = apiRoute;
-
-  // swap out [] for {} in route for AWS API Gateway compatibility
-  api.route(`ANY ${route.replace("[", "{").replace("]", "}")}`, {
-    bundle: `.aws-output/api/${id}`,
     handler: "index.handler",
     runtime: RUNTIME,
   });
