+++ infra/static-site.ts
